requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncpg>=0.29.0
python-dotenv==1.0.0
greenlet>=2.0.0
openai>=1.10.0
anthropic>=0.18.0
dateparser>=1.2.0
apscheduler>=3.10.0
numpy>=1.24.0
pytz>=2023.3
setuptools>=65.0.0
wheel>=0.40.0
python-dateutil>=2.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

